# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoJigGenerator API",
    "version": "baseline"
  },
  "paths": {
...
    "/api/works/{work_id}/promote/jig": {
      "post": {
        "tags": [
          "works"
        ],
        "summary": "Promote Jig",
        "description": "지그 생성 작업 하나를 지그 카탈로그에 올린다. 제품이 부품에 없으면 함께 올린다.",
        "operationId": "promote_jig_api_works__work_id__promote_jig_post",
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteJigRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoteJigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jigs": {
      "get": {
        "tags": [
          "jigs"
        ],
        "summary": "List Jigs",
        "operationId": "list_jigs_api_jigs_get",
        "parameters": [
          {
            "name": "part_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Part Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JigSummaryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jigs/{jig_id}": {
      "get": {
        "tags": [
          "jigs"
        ],
        "summary": "Get Jig",
        "operationId": "get_jig_api_jigs__jig_id__get",
        "parameters": [
          {
            "name": "jig_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Jig Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "jigs"
        ],
        "summary": "Update Jig",
        "operationId": "update_jig_api_jigs__jig_id__patch",
        "parameters": [
          {
            "name": "jig_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Jig Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JigUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jigs"
        ],
        "summary": "Delete Jig",
        "operationId": "delete_jig_api_jigs__jig_id__delete",
        "parameters": [
          {
            "name": "jig_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Jig Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jigs/{jig_id}/versions": {
      "get": {
        "tags": [
          "jigs"
        ],
        "summary": "List Versions",
        "operationId": "list_versions_api_jigs__jig_id__versions_get",
        "parameters": [
          {
            "name": "jig_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Jig Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JigVersionOut"
                  },
                  "title": "Response List Versions Api Jigs  Jig Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cad/recipe/check": {
      "post": {
        "tags": [
          "cad"
        ],
        "summary": "Recipe Check",
        "description": "모양만 본다(만들지 않는다). 편집기가 칸을 고칠 때마다 부른다.",
        "operationId": "recipe_check_api_cad_recipe_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeProblemsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cad/recipe/info": {
      "post": {
        "tags": [
          "cad"
        ],
        "summary": "Recipe Info",
        "description": "만들어 본 요약(크기 · 부피 · 노드별 정보). 실패하면 어느 노드가 왜인지.",
        "operationId": "recipe_info_api_cad_recipe_info_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeInfoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cad/recipe/preview": {
      "post": {
        "tags": [
          "cad"
        ],
        "summary": "Recipe Preview",
        "description": "미리보기 glTF.",
        "operationId": "recipe_preview_api_cad_recipe_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cad/recipe/step": {
      "post": {
        "tags": [
          "cad"
        ],
        "summary": "Recipe Step",
        "description": "저장하지 않고 STEP 만 받는다 — 한 번 쓰고 말 도형.",
        "operationId": "recipe_step_api_cad_recipe_step_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "work_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Id"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "work_id",
          "parent_id",
          "kind",
          "filename",
          "content_type",
          "size_bytes",
          "created_at"
        ],
        "title": "ArtifactOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JigOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "work_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "part_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Name"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "version_count": {
            "type": "integer",
            "title": "Version Count"
          },
          "current": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JigVersionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "owner_id",
          "owner_name",
          "work_id",
          "part_id",
          "part_name",
          "current_version",
          "version_count",
          "current",
          "created_at",
          "updated_at"
        ],
        "title": "JigOut"
      },
...
      "JigSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "part_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Name"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "interference_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interference Ok"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "owner_name",
          "part_id",
          "part_name",
          "current_version",
          "interference_ok",
          "updated_at"
        ],
        "title": "JigSummaryOut"
      },
      "JigUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "JigUpdateRequest"
      },
      "JigVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "jig_id": {
            "type": "string",
            "format": "uuid",
            "title": "Jig Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "part_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Name"
          },
          "part_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Version"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "promoted_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promoted By Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "jig_id",
          "number",
          "job",
          "options",
          "summary",
          "part_id",
          "part_name",
          "part_version",
          "note",
          "promoted_by_name",
          "created_at"
        ],
        "title": "JigVersionOut"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "requested_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Id"
          },
          "requested_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Name"
          },
          "work_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Id"
          },
          "work_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Name"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          },
          "progress": {
            "items": {
              "$ref": "#/components/schemas/StageOut"
            },
            "type": "array",
            "title": "Progress"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactOut"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "requested_by_id",
          "requested_by_name",
          "work_id",
          "work_name",
          "options",
          "progress",
          "summary",
          "error",
          "artifacts",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobOut"
      },
...
      "Page_JigSummaryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JigSummaryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[JigSummaryOut]"
      },
...
      "PromoteJigOut": {
        "properties": {
          "jig_id": {
            "type": "string",
            "format": "uuid",
            "title": "Jig Id"
          },
          "jig_version": {
            "type": "integer",
            "title": "Jig Version"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "part_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Version"
          },
          "part_promoted_now": {
            "type": "boolean",
            "title": "Part Promoted Now"
          }
        },
        "type": "object",
        "required": [
          "jig_id",
          "jig_version",
          "part_id",
          "part_version",
          "part_promoted_now"
        ],
        "title": "PromoteJigOut"
      },
      "PromoteJigRequest": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          },
          "promote_product": {
            "type": "boolean",
            "title": "Promote Product",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "PromoteJigRequest",
        "description": "지그 생성 작업 하나를 지그로. 제품(그때의 형상 버전)이 부품에 없으면 함께 승격한다."
      },
...
      "RecipeInfoOut": {
        "properties": {
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "RecipeInfoOut",
        "description": "레시피를 실제로 만들어 본 결과(요약). 화면 미리보기가 glTF 와 함께 받는다."
      },
      "RecipeProblemsOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "problems": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Problems"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "problems"
        ],
        "title": "RecipeProblemsOut"
      },
      "RecipeRequest": {
        "properties": {
          "recipe": {
            "additionalProperties": true,
            "type": "object",
            "title": "Recipe"
          }
        },
        "type": "object",
        "required": [
          "recipe"
        ],
        "title": "RecipeRequest"
      },
...
      "StageOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "millis": {
            "type": "integer",
            "title": "Millis"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "name",
          "millis",
          "detail"
        ],
        "title": "StageOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}